+++ go.mod
@@ -1,17 +1,17 @@
-module main
+module Dirigent
 
 go 1.18
 
 require (
 	github.com/andybalholm/brotli v1.0.4 // indirect
+	github.com/antoniodipinto/ikisocket v0.0.0-20220426120810-e8049cc9e56b // indirect
 	github.com/fasthttp/websocket v1.5.0 // indirect
-	github.com/gofiber/fiber/v2 v2.31.0 // indirect
-	github.com/gofiber/websocket/v2 v2.0.20 // indirect
-	github.com/gorilla/websocket v1.5.0 // indirect
-	github.com/klauspost/compress v1.15.1 // indirect
-	github.com/savsgio/gotils v0.0.0-20220401102855-e56b59f40436 // indirect
+	github.com/gofiber/fiber/v2 v2.34.0 // indirect
+	github.com/gofiber/websocket/v2 v2.0.22 // indirect
+	github.com/klauspost/compress v1.15.5 // indirect
+	github.com/savsgio/gotils v0.0.0-20220530130905-52f3993e8d6d // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasthttp v1.35.0 // indirect
+	github.com/valyala/fasthttp v1.37.0 // indirect
 	github.com/valyala/tcplisten v1.0.0 // indirect
-	golang.org/x/sys v0.0.0-20220412071739-889880a91fd5 // indirect
+	golang.org/x/sys v0.0.0-20220520151302-bc2c85ada10a // indirect
 )
