--- static/client.html
@@ -1,12 +1,0 @@
-<html>
-<head>
-    <title>WebSocket Demo</title>
-    <script src="http://dirigent.hanskazan.space/dep/script.js"></script>
-    <link rel="stylesheet" href="http://dirigent.hanskazan.space/dep/style.css">
-</head>
-<body>
-    <div id="container">
-        <p id="displayState">Nog niet verbonden... <noscript>Omdat je browser Javascript blokkeert!</noscript></p>
-    </div>
-</body>
-</html>
